+++ tapo-py/src/api/child_devices.rs
@@ -1,4 +1,5 @@
 mod ke100_handler;
+mod power_strip_plug_energy_monitoring_handler;
 mod power_strip_plug_handler;
 mod s200b_handler;
 mod t100_handler;
@@ -7,6 +8,7 @@ mod t300_handler;
 mod t31x_handler;
 
 pub use ke100_handler::*;
+pub use power_strip_plug_energy_monitoring_handler::*;
 pub use power_strip_plug_handler::*;
 pub use s200b_handler::*;
 pub use t31x_handler::*;