+++ tapo/src/api/child_devices/power_strip_plug_energy_monitoring_handler.rs
@@ -5,31 +5,31 @@ use tokio::sync::RwLock;
 use crate::api::ApiClient;
 use crate::error::{Error, TapoResponseError};
 use crate::requests::{EmptyParams, GenericSetDeviceInfoParams, TapoParams, TapoRequest};
-use crate::responses::{DecodableResultExt, PowerStripPlugResult};
+use crate::responses::{DecodableResultExt, PowerStripPlugEnergyMonitoringResult};
 
-/// Handler for the [P300](https://www.tp-link.com/en/search/?q=P300) and
-/// [P306](https://www.tp-link.com/us/search/?q=P306) child plugs.
-pub struct PowerStripPlugHandler {
+/// Handler for the [P304M](https://www.tp-link.com/uk/search/?q=P304M) and
+/// [P316M](https://www.tp-link.com/us/search/?q=P316M) child plugs.
+pub struct PowerStripPlugEnergyMonitoringHandler {
     client: Arc<RwLock<ApiClient>>,
     device_id: String,
 }
 
-impl PowerStripPlugHandler {
+impl PowerStripPlugEnergyMonitoringHandler {
     pub(crate) fn new(client: Arc<RwLock<ApiClient>>, device_id: String) -> Self {
         Self { client, device_id }
     }
 
-    /// Returns *device info* as [`PowerStripPlugResult`].
+    /// Returns *device info* as [`PowerStripPlugEnergyMonitoringResult`].
     /// It is not guaranteed to contain all the properties returned from the Tapo API.
     /// If the deserialization fails, or if a property that you care about it's not present,
-    /// try [`PowerStripPlugHandler::get_device_info_json`].
-    pub async fn get_device_info(&self) -> Result<PowerStripPlugResult, Error> {
+    /// try [`PowerStripPlugEnergyMonitoringHandler::get_device_info_json`].
+    pub async fn get_device_info(&self) -> Result<PowerStripPlugEnergyMonitoringResult, Error> {
         let request = TapoRequest::GetDeviceInfo(TapoParams::new(EmptyParams));
 
         self.client
             .read()
             .await
-            .control_child::<PowerStripPlugResult>(self.device_id.clone(), request)
+            .control_child::<PowerStripPlugEnergyMonitoringResult>(self.device_id.clone(), request)
             .await?
             .ok_or_else(|| Error::Tapo(TapoResponseError::EmptyResult))
             .map(|result| result.decode())?