+++ tapo/src/api/plug.rs
@@ -1,0 +1,9 @@
+/// Power strip plug.
+pub enum Plug {
+    ///  By Device ID.
+    ByDeviceId(String),
+    /// By Nickname.
+    ByNickname(String),
+    /// By Position.
+    ByPosition(u8),
+}