+++ tapo/src/responses/child_device_list_power_strip_result.rs
@@ -1,3 +1,7 @@
+mod auto_off_status;
+mod power_strip_plug_energy_monitoring_result;
 mod power_strip_plug_result;
 
+pub use auto_off_status::*;
+pub use power_strip_plug_energy_monitoring_result::*;
 pub use power_strip_plug_result::*;