+++ tapo/src/responses/child_device_list_power_strip_result/auto_off_status.rs
@@ -1,0 +1,11 @@
+use serde::{Deserialize, Serialize};
+
+/// Auto Off Status.
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+#[cfg_attr(feature = "python", pyo3::prelude::pyclass(get_all, eq, eq_int))]
+#[allow(missing_docs)]
+pub enum AutoOffStatus {
+    On,
+    Off,
+}