+++ tapo/src/responses/device_info_result/power_strip.rs
@@ -3,7 +3,7 @@ use serde::{Deserialize, Serialize};
 use crate::error::Error;
 use crate::responses::{DecodableResultExt, TapoResponseExt, decode_value};
 
-/// Device info of Tapo P300, P304M and P316M. Superset of [`crate::responses::DeviceInfoGenericResult`].
+/// Device info of Tapo P300, P304M, P306 and P316M. Superset of [`crate::responses::DeviceInfoGenericResult`].
 #[derive(Debug, Clone, Serialize, Deserialize)]
 #[cfg_attr(feature = "python", pyo3::prelude::pyclass(get_all))]
 #[allow(missing_docs)]
@@ -11,27 +11,27 @@ pub struct DeviceInfoPowerStripResult {
     //
     // Inherited from DeviceInfoGenericResult
     //
+    pub avatar: String,
     pub device_id: String,
-    pub r#type: String,
-    pub model: String,
-    pub hw_id: String,
-    pub hw_ver: String,
     pub fw_id: String,
     pub fw_ver: String,
-    pub oem_id: String,
-    pub mac: String,
-    pub ip: String,
-    pub ssid: String,
-    pub signal_level: u8,
-    pub rssi: i16,
-    pub specs: String,
-    pub lang: String,
-    pub avatar: String,
     pub has_set_location_info: bool,
-    pub region: Option<String>,
+    pub hw_id: String,
+    pub hw_ver: String,
+    pub ip: String,
+    pub lang: String,
     pub latitude: Option<i64>,
     pub longitude: Option<i64>,
-    pub time_diff: Option<i64>,
+    pub mac: String,
+    pub model: String,
+    pub oem_id: String,
+    pub region: Option<String>,
+    pub rssi: i16,
+    pub signal_level: u8,
+    pub specs: String,
+    pub ssid: String,
+    pub time_diff: i64,
+    pub r#type: String,
 }
 
 #[cfg(feature = "python")]